refactor(communications): move from core, email to communications
