+++ frappe/communications/doctype/communication/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (c) 2015, Frappe Technologies Pvt. Ltd. and Contributors
+# License: MIT. See LICENSE
