+++ frappe/communications/doctype/communication/communication.py
@@ -12,10 +12,10 @@ from frappe import _
 from frappe.automation.doctype.assignment_rule.assignment_rule import (
 	apply as apply_assignment_rule,
 )
+from frappe.communications.doctype.communication.email import validate_email
+from frappe.communications.doctype.communication.mixins import CommunicationEmailMixin
 from frappe.contacts.doctype.contact.contact import get_contact_name
 from frappe.core.doctype.comment.comment import update_comment_in_doc
-from frappe.core.doctype.communication.email import validate_email
-from frappe.core.doctype.communication.mixins import CommunicationEmailMixin
 from frappe.core.utils import get_parent_doc
 from frappe.model.document import Document
 from frappe.utils import (
@@ -38,7 +38,7 @@ class Communication(Document, CommunicationEmailMixin):
 	from typing import TYPE_CHECKING
 
 	if TYPE_CHECKING:
-		from frappe.core.doctype.communication_link.communication_link import CommunicationLink
+		from frappe.communications.doctype.communication_link.communication_link import CommunicationLink
 		from frappe.types import DF
 
 		_user_tags: DF.Data | None
