+++ frappe/communications/doctype/communication_link/communication_link.json
@@ -36,7 +36,7 @@
  "istable": 1,
  "modified": "2019-05-21 09:47:23.043960",
  "modified_by": "Administrator",
- "module": "Core",
+ "module": "Communications",
  "name": "Communication Link",
  "owner": "Administrator",
  "permissions": [],
