+++ frappe/communications/doctype/notification/notification.json
@@ -293,7 +293,7 @@
  "links": [],
  "modified": "2023-11-17 08:48:25.616203",
  "modified_by": "Administrator",
- "module": "Email",
+ "module": "Communications",
  "name": "Notification",
  "naming_rule": "Set by user",
  "owner": "Administrator",
