+++ frappe/communications/doctype/notification_recipient/notification_recipient.json
@@ -51,7 +51,7 @@
  "links": [],
  "modified": "2020-09-01 17:40:27.289105",
  "modified_by": "Administrator",
- "module": "Email",
+ "module": "Communications",
  "name": "Notification Recipient",
  "owner": "Administrator",
  "permissions": [],
